requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
black==20.8b1
click==7.1.2
dataclasses==0.8
Django==3.1.5
django-extensions==3.1.0
djangorestframework==3.12.2
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pathspec==0.8.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.4.0